Functions

+++ src/vm/debug/debug.h
@@ -3,7 +3,7 @@
 
 #include "vm/compiler/chunk.h"
 
-int32_t bluDisassembleInstruction(bluChunk* chunk, size_t offset);
+int32_t bluDisassembleInstruction(bluChunk* chunk, int32_t offset);
 void bluDisassembleChunk(bluChunk* chunk);
 
 #endif
